GITechDemo changes: *Some RSM optimizations in the form of a quarter-resolution indirect light accumulation buffer, which gets additively color blended over the full-res light accumulation buffer using edge-aware bilateral upscaling *Added a new PCF filter type with a 4x4 Poisson Disk kernel rotated 0, 90, 180 and 270 degrees around its center (although more bandwidth intensive, it further improved the shadow's jagged edges) *Created a script for automating the build process
LibRenderer changes:
*Small bug fixes

+++ GITechDemo/Code/Libraries/LibRenderer/Common/Texture.cpp
@@ -146,9 +146,9 @@ void Texture::ComputeTextureProperties(const Vec3i dimensions)
 	if ((dimensions[0] <= 0 || dimensions[1] <= 0) && (IsRenderTarget() || IsDepthStencil()))
 		m_bIsDynamicRT = true;
 
-	unsigned int width	= !m_bIsDynamicRT ? dimensions[0] : Renderer::GetInstance()->GetBackBufferSize()[0] * m_fWidthRatio;
-	unsigned int height	= !m_bIsDynamicRT ? dimensions[1] : Renderer::GetInstance()->GetBackBufferSize()[1] * m_fHeightRatio;
-	unsigned int depth	= !m_bIsDynamicRT ? dimensions[2] : 1;
+	unsigned int width	= !m_bIsDynamicRT ? dimensions[0] : (unsigned int)((float)Renderer::GetInstance()->GetBackBufferSize()[0] * m_fWidthRatio);
+	unsigned int height	= !m_bIsDynamicRT ? dimensions[1] : (unsigned int)((float)Renderer::GetInstance()->GetBackBufferSize()[1] * m_fHeightRatio);
+	unsigned int depth	= !m_bIsDynamicRT ? dimensions[2] : 1u;
 	if (IsCompressed())
 	{
 		// Calculate the smallest number divisible by 4 which is >= width/height
